+++ deepclone.stub.php
@@ -23,9 +23,15 @@ namespace {
      */
     const DEEPCLONE_HYDRATE_NO_LAZY_INIT = UNKNOWN;
 
+    /**
+     * @var int
+     * @cvalue DEEPCLONE_HYDRATE_MANGLED_VARS
+     */
+    const DEEPCLONE_HYDRATE_MANGLED_VARS = UNKNOWN;
+
     function deepclone_to_array(mixed $value, ?array $allowed_classes = null): array {}
 
     function deepclone_from_array(array $data, ?array $allowed_classes = null): mixed {}
 
-    function deepclone_hydrate(object|string $object_or_class, array $scoped_vars = [], array $mangled_vars = [], int $flags = 0): object {}
+    function deepclone_hydrate(object|string $object_or_class, array $vars = [], int $flags = 0): object {}
 }